// web/src/shared/api/types.ts
export const OpenIdProviderKind = {
  Custom: 'Custom',
  Google: 'Google',
  Microsoft: 'Microsoft',
  Okta: 'Okta',
  JumpCloud: 'JumpCloud',
  Zitadel: 'Zitadel',
} as const;

export type OpenIdProviderKindValue =
  (typeof OpenIdProviderKind)[keyof typeof OpenIdProviderKind];